IQFSV1/requirements.txt
streamlit>=1.33
anthropic>=0.39
apify-client>=1.7
python-dotenv>=1.0
pypdf>=4.2
httpx>=0.27
